requirements.txt
requests==2.25.1
PyYAML==5.4.1
openshift==0.11.2
configparser==5.0.1
Jinja2==2.11.3
pydantic==1.8.2
